requirements.txt
enlighten
fonttools
matplotlib
numpy
Pillow
pycocotools
torch
torchvision
tensorboard
